Hide paging bar if only one page presents
There are different opinions whether the paging bar should be shown if only only page presents
(https://ux.stackexchange.com/questions/28401/should-the-pagination-widget-be-shown-when-there-is-only-1-page).

Seems like for current application it's better to hide bar in this case.
Usually the number of movies on the page is small and in most cases there will be just one page.