Remove top paging bar
Usually it's recommended to display paging bar at the top and at the bottom of the page.
However we remove top paging bar. It overloads top part of the page (navigation pane, add movie form, paging bar).
Another reason is that in most cases the user will not randomly jump between the pages.
We expect that the user will thoughtfully study the content, page by page.