Add confirmation page for adding of new movies
This commit also removes jump to just added movie and fade-in effect.
Such jump was used previously, so that user can examine the info parsed for just added movie.
Now the user examines the parsing on the confirmation page. And it's not trivial to implement the jump after paging was introduced.