Allow default admin login from any host
Previously we allowed login for default admin only from localhost. It worked fine for deployments to VMs.
However with deployment to Azure App service, we can not access the application from localhost.
That is why we remove localhost restriction for default admin.

Here is the flow for initial application launch:

1. First login is performed under default admin.
2. Another user with own login and password is created.
3. New user is granted with administrator role.
4. Default admin is deleted.