Enhance override of connection string in IT
We want to deploy application as Azure App Service.
The database will be hosted at remote MongoDB instance, which requires authorized access.
To make development and testing environments closer to production, we also configure local MongoDB for authorized access.

That is why we want to allow overriding of connection string via environment variable, so that sensitive credentials are stored safely.

The tricky moment is to separate connection strings for IT and development databases in local environment (so that IT do not override data added manually).
To handle this we added `connectionStringName` setting which defines the source name for connection string.