Change default logs path
Previously the deployment specific path was used (The relative directory path which is specific for hosting in IIS on VM).
This does not work for other deployments, e.g. Docker or Azure App service.

So, we change the path to reasonable default - directory "logs".
It should be adjusted for specific deployments with as needed, e.g. via environment variables.